seems most of the linux bugs were due to me not initialising the windows co-ordinates  stuff (rotXY pos etc) in NGLScene. It seems that clang / msvc compilers seems to set them to zero g++ and clang under linux do not

+++ Circle/src/NGLScene.cpp
@@ -53,8 +53,8 @@ void NGLScene::setupSim()
   for (size_t i = 0; i < 250; ++i)
   {
     m_sim->addAgent(200.0f *
-                  RVO::Vector2(std::cosf(i * 2.0f * M_PI / 250.0f),
-                               std::sinf(i * 2.0f * M_PI / 250.0f)));
+                  RVO::Vector2(std::cos(i * 2.0f * M_PI / 250.0f),
+                               std::sin(i * 2.0f * M_PI / 250.0f)));
     m_goals.push_back(-m_sim->getAgentPosition(i));
   }
 }
